* commit 'b9c5d606aab6c1a21e3785d155613c35c276eb18':
  Change android_dlextinfo flags to uint64_t

+++ libc/include/android/dlext.h
@@ -63,7 +63,7 @@ enum {
 };
 
 typedef struct {
-  int     flags;
+  uint64_t flags;
   void*   reserved_addr;
   size_t  reserved_size;
   int     relro_fd;
